+++ apps/docs/components/testimonials/TestimonialContainer.tsx
@@ -24,15 +24,15 @@ const TestimonialView: FC<Testimonial> = (testimonial) => {
         <div className="flex flex-col gap-3 rounded-lg border bg-card p-6 shadow transition-colors hover:bg-border">
           <div className="relative flex items-center gap-2">
             <Image
-              alt={"@" + testimonial.username + "'s twitter image"}
+              alt={`@${testimonial.username}'s twitter image`}
               loading="lazy"
               width="64"
               height="64"
               className="h-10 w-10 rounded-full border"
               src={testimonial.avatar}
             />
-            <p className="text-sm font-medium">{testimonial.username}</p>
-            <div className="absolute -top-1 -left-1 flex h-5 w-5 items-center justify-center rounded-full bg-background">
+            <p className="font-medium text-sm">{testimonial.username}</p>
+            <div className="-top-1 -left-1 absolute flex h-5 w-5 items-center justify-center rounded-full bg-background">
               <XLogo />
             </div>
           </div>