+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,17 @@
 # assistant-ui
 
+## 0.0.67
+
+### Patch Changes
+
+- ec662cd: chore: update dependencies
+
+## 0.0.66
+
+### Patch Changes
+
+- 9d50ed9: refactor assistant-ui cli
+
 ## 0.0.65
 
 ### Patch Changes